refactor TestCase to be placed in src/masonite/tests
