+++ tests/core/authentication/test_authentication2.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 from src.masonite.foundation import Application
 import os
 from masoniteorm.models import Model
