+++ tests/core/middleware/test_encrypt_cookies.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 from src.masonite.middleware import EncryptCookies
