+++ tests/core/request/test_http_requests.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 
 
 class TestHttpRequests(TestCase):
