+++ tests/features/cache/test_file_cache.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 import os
 import time
