+++ tests/features/queues/test_async_driver.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 from src.masonite.queues import Queueable
 import os
 import time
