+++ tests/features/session/test_cookie_session.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 
 
 class TestCookieSession(TestCase):
