+++ tests/integrations/templates/welcome.html
@@ -11,7 +11,7 @@ Hello. Setup chat connection here
 </head>
 <body>
     <h1>Welcome</h1>
-    
+    <img src="favicon.ico" alt="img test">
     <script>
         const pusher = new Pusher("", {
             authEndpoint: '/broadcasting/authorize',
