+++ tests/tests/test_transactions.py
@@ -1,4 +1,4 @@
-from tests import TestCase
+from src.masonite.tests import TestCase
 from tests.integrations.app.User import User
 from src.masonite.tests import DatabaseTransactions
